refactor: product 패키지 리팩토링
